backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.1
python-slugify==8.0.4
Markdown==3.6
gunicorn==22.0.0
